+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 Changelog
 =========
 
+## 1.0.1 (2020-03-21)
+ * Check file not exists or generate unique filename
+ * Update README.md
+
 ## 1.0.0 (2020-03-20)
  * Added bulk actions for media items
  * CRUD for media items/categories, translations